[pull] master from lightningnetwork:master

+++ walletunlocker/service.go
@@ -16,6 +16,13 @@ import (
 	"github.com/lightningnetwork/lnd/lnrpc"
 	"github.com/lightningnetwork/lnd/lnwallet"
 	"github.com/lightningnetwork/lnd/lnwallet/btcwallet"
+	"github.com/lightningnetwork/lnd/macaroons"
+)
+
+var (
+	// ErrUnlockTimeout signals that we did not get the expected unlock
+	// message before the timeout occurred.
+	ErrUnlockTimeout = errors.New("got no unlock message before timeout")
 )
 
 // ChannelsToRecover wraps any set of packed (serialized+encrypted) channel
@@ -54,6 +61,11 @@ type WalletInitMsg struct {
 	// ChanBackups a set of static channel backups that should be received
 	// after the wallet has been initialized.
 	ChanBackups ChannelsToRecover
+
+	// StatelessInit signals that the user requested the daemon to be
+	// initialized stateless, which means no unencrypted macaroons should be
+	// written to disk.
+	StatelessInit bool
 }
 
 // WalletUnlockMsg is a message sent by the UnlockerService when a user wishes
@@ -85,6 +97,11 @@ type WalletUnlockMsg struct {
 	// UnloadWallet is a function for unloading the wallet, which should
 	// be called on shutdown.
 	UnloadWallet func() error
+
+	// StatelessInit signals that the user requested the daemon to be
+	// initialized stateless, which means no unencrypted macaroons should be
+	// written to disk.
+	StatelessInit bool
 }
 
 // UnlockerService implements the WalletUnlocker service used to provide lnd
@@ -100,22 +117,46 @@ type UnlockerService struct {
 	// sent.
 	UnlockMsgs chan *WalletUnlockMsg
 
+	// MacResponseChan is the channel for sending back the admin macaroon to
+	// the WalletUnlocker service.
+	MacResponseChan chan []byte
+
 	chainDir       string
 	noFreelistSync bool
 	netParams      *chaincfg.Params
-	macaroonFiles  []string
+
+	// macaroonFiles is the path to the three generated macaroons with
+	// different access permissions. These might not exist in a stateless
+	// initialization of lnd.
+	macaroonFiles []string
+
+	// dbTimeout specifies the timeout value to use when opening the wallet
+	// database.
+	dbTimeout time.Duration
+
+	// resetWalletTransactions indicates that the wallet state should be
+	// reset on unlock to force a full chain rescan.
+	resetWalletTransactions bool
 }
 
 // New creates and returns a new UnlockerService.
 func New(chainDir string, params *chaincfg.Params, noFreelistSync bool,
-	macaroonFiles []string) *UnlockerService {
+	macaroonFiles []string, dbTimeout time.Duration,
+	resetWalletTransactions bool) *UnlockerService {
 
 	return &UnlockerService{
-		InitMsgs:      make(chan *WalletInitMsg, 1),
-		UnlockMsgs:    make(chan *WalletUnlockMsg, 1),
-		chainDir:      chainDir,
-		netParams:     params,
-		macaroonFiles: macaroonFiles,
+		InitMsgs:   make(chan *WalletInitMsg, 1),
+		UnlockMsgs: make(chan *WalletUnlockMsg, 1),
+
+		// Make sure we buffer the channel is buffered so the main lnd
+		// goroutine isn't blocking on writing to it.
+		MacResponseChan:         make(chan []byte, 1),
+		chainDir:                chainDir,
+		netParams:               params,
+		macaroonFiles:           macaroonFiles,
+		dbTimeout:               dbTimeout,
+		noFreelistSync:          noFreelistSync,
+		resetWalletTransactions: resetWalletTransactions,
 	}
 }
 
@@ -127,13 +168,15 @@ func New(chainDir string, params *chaincfg.Params, noFreelistSync bool,
 // Once the cipherseed is obtained and verified by the user, the InitWallet
 // method should be used to commit the newly generated seed, and create the
 // wallet.
-func (u *UnlockerService) GenSeed(ctx context.Context,
+func (u *UnlockerService) GenSeed(_ context.Context,
 	in *lnrpc.GenSeedRequest) (*lnrpc.GenSeedResponse, error) {
 
 	// Before we start, we'll ensure that the wallet hasn't already created
 	// so we don't show a *new* seed to the user if one already exists.
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	loader := wallet.NewLoader(u.netParams, netDir, u.noFreelistSync, 0)
+	loader := wallet.NewLoader(
+		u.netParams, netDir, u.noFreelistSync, u.dbTimeout, 0,
+	)
 	walletExists, err := loader.WalletExists()
 	if err != nil {
 		return nil, err
@@ -263,7 +306,8 @@ func (u *UnlockerService) InitWallet(ctx context.Context,
 	// wallet's files so we can check if the wallet already exists.
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
 	loader := wallet.NewLoader(
-		u.netParams, netDir, u.noFreelistSync, uint32(recoveryWindow),
+		u.netParams, netDir, u.noFreelistSync,
+		u.dbTimeout, uint32(recoveryWindow),
 	)
 
 	walletExists, err := loader.WalletExists()
@@ -297,6 +341,7 @@ func (u *UnlockerService) InitWallet(ctx context.Context,
 		Passphrase:     password,
 		WalletSeed:     cipherSeed,
 		RecoveryWindow: uint32(recoveryWindow),
+		StatelessInit:  in.StatelessInit,
 	}
 
 	// Before we return the unlock payload, we'll check if we can extract
@@ -306,9 +351,25 @@ func (u *UnlockerService) InitWallet(ctx context.Context,
 		initMsg.ChanBackups = *chansToRestore
 	}
 
-	u.InitMsgs <- initMsg
+	// Deliver the initialization message back to the main daemon.
+	select {
+	case u.InitMsgs <- initMsg:
+		// We need to read from the channel to let the daemon continue
+		// its work and to get the admin macaroon. Once the response
+		// arrives, we directly forward it to the client.
+		select {
+		case adminMac := <-u.MacResponseChan:
+			return &lnrpc.InitWalletResponse{
+				AdminMacaroon: adminMac,
+			}, nil
 
-	return &lnrpc.InitWalletResponse{}, nil
+		case <-ctx.Done():
+			return nil, ErrUnlockTimeout
+		}
+
+	case <-ctx.Done():
+		return nil, ErrUnlockTimeout
+	}
 }
 
 // UnlockWallet sends the password provided by the incoming UnlockWalletRequest
@@ -322,7 +383,8 @@ func (u *UnlockerService) UnlockWallet(ctx context.Context,
 
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
 	loader := wallet.NewLoader(
-		u.netParams, netDir, u.noFreelistSync, recoveryWindow,
+		u.netParams, netDir, u.noFreelistSync,
+		u.dbTimeout, recoveryWindow,
 	)
 
 	// Check if wallet already exists.
@@ -344,6 +406,37 @@ func (u *UnlockerService) UnlockWallet(ctx context.Context,
 		return nil, err
 	}
 
+	// The user requested to drop their whole wallet transaction state to
+	// force a full chain rescan for wallet addresses. Dropping the state
+	// only properly takes effect after opening the wallet. That's why we
+	// start, drop, stop and start again.
+	if u.resetWalletTransactions {
+		dropErr := wallet.DropTransactionHistory(
+			unlockedWallet.Database(), true,
+		)
+
+		// Even if dropping the history fails, we'll want to unload the
+		// wallet. If unloading fails, that error is probably more
+		// important to be returned to the user anyway.
+		if err := loader.UnloadWallet(); err != nil {
+			return nil, fmt.Errorf("could not unload "+
+				"wallet (tx history drop err: %v): %v", dropErr,
+				err)
+		}
+
+		// If dropping failed but unloading didn't, we'll still abort
+		// and inform the user.
+		if dropErr != nil {
+			return nil, dropErr
+		}
+
+		// All looks good, let's now open the wallet again.
+		unlockedWallet, err = loader.OpenExistingWallet(password, false)
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	// We successfully opened the wallet and pass the instance back to
 	// avoid it needing to be unlocked again.
 	walletUnlockMsg := &WalletUnlockMsg{
@@ -351,6 +444,7 @@ func (u *UnlockerService) UnlockWallet(ctx context.Context,
 		RecoveryWindow: recoveryWindow,
 		Wallet:         unlockedWallet,
 		UnloadWallet:   loader.UnloadWallet,
+		StatelessInit:  in.StatelessInit,
 	}
 
 	// Before we return the unlock payload, we'll check if we can extract
@@ -360,12 +454,25 @@ func (u *UnlockerService) UnlockWallet(ctx context.Context,
 		walletUnlockMsg.ChanBackups = *chansToRestore
 	}
 
-	// At this point we was able to open the existing wallet with the
+	// At this point we were able to open the existing wallet with the
 	// provided password. We send the password over the UnlockMsgs
 	// channel, such that it can be used by lnd to open the wallet.
-	u.UnlockMsgs <- walletUnlockMsg
+	select {
+	case u.UnlockMsgs <- walletUnlockMsg:
+		// We need to read from the channel to let the daemon continue
+		// its work. But we don't need the returned macaroon for this
+		// operation, so we read it but then discard it.
+		select {
+		case <-u.MacResponseChan:
+			return &lnrpc.UnlockWalletResponse{}, nil
 
-	return &lnrpc.UnlockWalletResponse{}, nil
+		case <-ctx.Done():
+			return nil, ErrUnlockTimeout
+		}
+
+	case <-ctx.Done():
+		return nil, ErrUnlockTimeout
+	}
 }
 
 // ChangePassword changes the password of the wallet and sends the new password
@@ -375,7 +482,9 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 	in *lnrpc.ChangePasswordRequest) (*lnrpc.ChangePasswordResponse, error) {
 
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	loader := wallet.NewLoader(u.netParams, netDir, u.noFreelistSync, 0)
+	loader := wallet.NewLoader(
+		u.netParams, netDir, u.noFreelistSync, u.dbTimeout, 0,
+	)
 
 	// First, we'll make sure the wallet exists for the specific chain and
 	// network.
@@ -408,18 +517,32 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 	if err != nil {
 		return nil, err
 	}
-	// Unload the wallet to allow lnd to open it later on.
-	defer loader.UnloadWallet()
 
-	// Since the macaroon database is also encrypted with the wallet's
-	// password, we'll remove all of the macaroon files so that they're
-	// re-generated at startup using the new password. We'll make sure to do
-	// this after unlocking the wallet to ensure macaroon files don't get
-	// deleted with incorrect password attempts.
-	for _, file := range u.macaroonFiles {
-		err := os.Remove(file)
-		if err != nil && !os.IsNotExist(err) {
-			return nil, err
+	// Now that we've opened the wallet, we need to close it in case of an
+	// error. But not if we succeed, then the caller must close it.
+	orderlyReturn := false
+	defer func() {
+		if !orderlyReturn {
+			_ = loader.UnloadWallet()
+		}
+	}()
+
+	// Before we actually change the password, we need to check if all flags
+	// were set correctly. The content of the previously generated macaroon
+	// files will become invalid after we generate a new root key. So we try
+	// to delete them here and they will be recreated during normal startup
+	// later. If they are missing, this is only an error if the
+	// stateless_init flag was not set.
+	if in.NewMacaroonRootKey || in.StatelessInit {
+		for _, file := range u.macaroonFiles {
+			err := os.Remove(file)
+			if err != nil && !in.StatelessInit {
+				return nil, fmt.Errorf("could not remove "+
+					"macaroon file: %v. if the wallet "+
+					"was initialized stateless please "+
+					"add the --stateless_init "+
+					"flag", err)
+			}
 		}
 	}
 
@@ -434,11 +557,86 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 			"%v", err)
 	}
 
+	// The next step is to load the macaroon database, change the password
+	// then close it again.
+	// Attempt to open the macaroon DB, unlock it and then change
+	// the passphrase.
+	macaroonService, err := macaroons.NewService(
+		netDir, "lnd", in.StatelessInit, u.dbTimeout,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	err = macaroonService.CreateUnlock(&privatePw)
+	if err != nil {
+		closeErr := macaroonService.Close()
+		if closeErr != nil {
+			return nil, fmt.Errorf("could not create unlock: %v "+
+				"--> follow-up error when closing: %v", err,
+				closeErr)
+		}
+		return nil, err
+	}
+	err = macaroonService.ChangePassword(privatePw, in.NewPassword)
+	if err != nil {
+		closeErr := macaroonService.Close()
+		if closeErr != nil {
+			return nil, fmt.Errorf("could not change password: %v "+
+				"--> follow-up error when closing: %v", err,
+				closeErr)
+		}
+		return nil, err
+	}
+
+	// If requested by the user, attempt to replace the existing
+	// macaroon root key with a new one.
+	if in.NewMacaroonRootKey {
+		err = macaroonService.GenerateNewRootKey()
+		if err != nil {
+			closeErr := macaroonService.Close()
+			if closeErr != nil {
+				return nil, fmt.Errorf("could not generate "+
+					"new root key: %v --> follow-up error "+
+					"when closing: %v", err, closeErr)
+			}
+			return nil, err
+		}
+	}
+
+	err = macaroonService.Close()
+	if err != nil {
+		return nil, fmt.Errorf("could not close macaroon service: %v",
+			err)
+	}
+
 	// Finally, send the new password across the UnlockPasswords channel to
 	// automatically unlock the wallet.
-	u.UnlockMsgs <- &WalletUnlockMsg{Passphrase: in.NewPassword}
+	walletUnlockMsg := &WalletUnlockMsg{
+		Passphrase:    in.NewPassword,
+		Wallet:        w,
+		StatelessInit: in.StatelessInit,
+		UnloadWallet:  loader.UnloadWallet,
+	}
+	select {
+	case u.UnlockMsgs <- walletUnlockMsg:
+		// We need to read from the channel to let the daemon continue
+		// its work and to get the admin macaroon. Once the response
+		// arrives, we directly forward it to the client.
+		orderlyReturn = true
+		select {
+		case adminMac := <-u.MacResponseChan:
+			return &lnrpc.ChangePasswordResponse{
+				AdminMacaroon: adminMac,
+			}, nil
 
-	return &lnrpc.ChangePasswordResponse{}, nil
+		case <-ctx.Done():
+			return nil, ErrUnlockTimeout
+		}
+
+	case <-ctx.Done():
+		return nil, ErrUnlockTimeout
+	}
 }
 
 // ValidatePassword assures the password meets all of our constraints.
